shaderthing - iResolution is now a vec2 instead of ivec2; testing a different approach to feeding layer framebuffer color attachments to other layers as uniforms (passing the writeOnlyFramebuffer color attachment instead of the readOnly one, except if the receiving layer is the same as the source layer, for obvious reasons) | vir - preparing groundwork from eventual (if possible), switch from using a GL_RGBA(8UI) internal texture (inclusive of framebuffer color attachments) format to a GL_RGBA32F for greater accuracy when using layers to store raw data; preliminary testing indicates that it should be possibile, but the OpenGLKMeansColorQuantizer will likely have to be re-adapted to this

+++ shaderthing/src/data/coderepository_gui.cpp
@@ -127,7 +127,7 @@ out vec4 fragColor;
 in vec2 pos;
 in vec2 txc;
 
-uniform ivec2 iResolution;
+uniform vec2 iResolution;
 uniform sampler2D iTexture0;
 
 void make_kernel(inout vec4 n[9], sampler2D tex, vec2 coord)
